Rename match origin to match host
