+++ live.gen.ts
@@ -22,9 +22,9 @@ import * as $$$$$$$$0 from "./sections/UseSlot.tsx";
 import * as $$$$$$$$1 from "./sections/Slot.tsx";
 import * as $$$$$$$$2 from "./sections/PageInclude.tsx";
 import * as $$$$$$$$$0 from "./matchers/MatchDate.ts";
-import * as $$$$$$$$$1 from "./matchers/MatchOrigin.ts";
-import * as $$$$$$$$$2 from "./matchers/MatchUserAgent.ts";
-import * as $$$$$$$$$3 from "./matchers/MatchSite.ts";
+import * as $$$$$$$$$1 from "./matchers/MatchUserAgent.ts";
+import * as $$$$$$$$$2 from "./matchers/MatchSite.ts";
+import * as $$$$$$$$$3 from "./matchers/MatchHost.ts";
 import * as $$$$$$$$$4 from "./matchers/MatchMulti.ts";
 import * as $$$$$$$$$5 from "./matchers/MatchRandom.ts";
 import * as $$$$$$$$$6 from "./matchers/MatchEnvironment.ts";
@@ -64,11 +64,11 @@ const manifest = {
     "$live/matchers/MatchAlways.ts": $$$$$$$$$7,
     "$live/matchers/MatchDate.ts": $$$$$$$$$0,
     "$live/matchers/MatchEnvironment.ts": $$$$$$$$$6,
+    "$live/matchers/MatchHost.ts": $$$$$$$$$3,
     "$live/matchers/MatchMulti.ts": $$$$$$$$$4,
-    "$live/matchers/MatchOrigin.ts": $$$$$$$$$1,
     "$live/matchers/MatchRandom.ts": $$$$$$$$$5,
-    "$live/matchers/MatchSite.ts": $$$$$$$$$3,
-    "$live/matchers/MatchUserAgent.ts": $$$$$$$$$2,
+    "$live/matchers/MatchSite.ts": $$$$$$$$$2,
+    "$live/matchers/MatchUserAgent.ts": $$$$$$$$$1,
   },
   "flags": {
     "$live/flags/audience.ts": $$$$$$$$$$0,
